printf putvoid experiment

+++ 1.1__ft-printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: jniedens <jniedens@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/30 11:49:38 by jniedens          #+#    #+#             */
-/*   Updated: 2023/01/02 16:03:37 by jniedens         ###   ########.fr       */
+/*   Updated: 2023/01/02 16:08:45 by jniedens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ static int	ft_check(char c, va_list args)
 	else if (c == 'X')
 		count += (ft_puthexup(va_arg(args, unsigned int), 0));
 	else if (c == 'p')
-		count += (ft_putpointer(va_arg(args, unsigned long)));
+		count += (ft_putpointer(va_arg(args, unsigned long long)));
 	else if (c == '%')
 	{
 		count++;
